add option to (left) join collection while reading nodes
Example use case:
For each node in nodes collection we want to store scores for some property.
The scores are computed by some model. We can have different versions of that
model. Instead of storing each version as attributes in the nodes collection
using different names we store them as separate collections.
When loading the nodes we (left) join the chosen collection onto the nodes
using the node ID.